Move controllers from lib to a new controllers folder
