+++ main.lua
@@ -7,16 +7,18 @@ lovetoys.initialize(
     globals = true
   }
 )
-Stack = require "lib/Stack"
 
+Stack = require "lib/Stack"
 Utils = require("lib/Utils")
-SoundController = require("lib/SoundController")
-MenuController = require("lib/MenuController")
-WaveController = require("lib/WaveController")
-StatisticController = require("lib/StatisticController")
-ActionsController = require("lib/ActionsController")
-LogController = require("lib/LogController")
 GameState = require "lib/GameState"
+
+-- controllers
+ActionsController = require("controllers/ActionsController")
+LogController = require("controllers/LogController")
+MenuController = require("controllers/MenuController")
+SoundController = require("controllers/SoundController")
+StatisticController = require("controllers/StatisticController")
+WaveController = require("controllers/WaveController")
 
 --- components
 require "components/AI"
